% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_rows_true}
\alias{all_rows_true}
\title{Function to check if all values in a row are true across an entire dataframe.}
\usage{
all_rows_true(input_tibble, col_name)
}
\description{
Function to check if all values in a row are true across an entire dataframe.
}
