% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_all_data}
\alias{check_all_data}
\title{Function to check all data requirements.
Simply iterates over the requirements and applies the changes
to the given id tibble based on the data-checking functions.}
\usage{
check_all_data(
  ids,
  required_list,
  modality = NA,
  col_suffix = NA,
  my_required = NA
)
}
\description{
Function to check all data requirements.
Simply iterates over the requirements and applies the changes
to the given id tibble based on the data-checking functions.
}
