% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_id_path_data}
\alias{check_id_path_data}
\title{Function to check that a subject has data for a set of ids.}
\usage{
check_id_path_data(ids, requirement)
}
\arguments{
\item{ids}{is a tibble with a column named "id".}

\item{requirements}{is a list as described in the description.}
}
\value{
a tibble with a variable that if the data exists for each id.
}
\description{
Requirements are min, max, grep, keywords, exclude, and
file_count. Min and max are the min and max number of matches
for a particular data check (set to -1 if not checking). The
grep argument is a grep string to match over. Keywords and
exclude are sub strings that must be included or must be ignored
in the grep match, respectively. Finally, file_count is the number
of file if the result is a directory (set to 0 if it is a file).
}
