% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DNPLsetup}
\alias{DNPLsetup}
\title{Function to be run to setup a base configuration for DNPL.}
\usage{
DNPLsetup(github_uname, github_token)
}
\arguments{
\item{github_uname}{Your GitHub username.}

\item{github_token}{Your GitHub access token.}
}
\value{
FALSE if not set, TRUE if set.
}
\description{
This function sets up the user's GitHub credentials and then
uses those credentials to configure Rclone and DataTracker.
This function is specific to the DNPL at UPMC.
}
\examples{
set_github_creds(github_uname=<username>, github_token=<token>)
}
