% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_datatracker_cfg}
\alias{fetch_datatracker_cfg}
\title{Function to fetch and set the datatracker config files.}
\usage{
fetch_datatracker_cfg(repo, path, gh_root = "", save_to = NA, set_lab = NA)
}
\arguments{
\item{repo}{Is the name of the repository.}

\item{path}{Is the path in the repo to the file.}

\item{gh_root}{If set to "" will try to access the file from your saved}

\item{save_to}{Is the path to save the datatracker config to.}

\item{set_lab}{Is if you are setting up a new lab/project.}
}
\description{
Function for getting a config file from GitHub. They can be given a specific
path to be saved to, or set as the user's default lab configuration.
}
