% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_github_data}
\alias{fetch_github_data}
\title{Function to fetch files from GitHub.}
\usage{
fetch_github_data(repo, path, gh_root = "")
}
\arguments{
\item{repo}{Is the name of the repository.}

\item{path}{Is the path in the repo to the file.}

\item{gh_root}{If set to "" will try to access the file from your saved
GitHub username. If a group is given, will attempt to fetch from that another
GitHub account or organization.}
}
\value{
The file as raw text from a GitHub repo.
}
\description{
Modified from:
https://gist.github.com/ritchieking/5de10cde6b46f3536967a908fe806b5f
}
\examples{
raw_file <- fetch_github_data(repo="Lab_Configs",
  path="datatracker/lab_cfg.json",
  gh_root="DecisionNeurosciencePsychopathology")
}
