% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_rclone_cfg}
\alias{fetch_rclone_cfg}
\title{Function to get and set an rclone config file.}
\usage{
fetch_rclone_cfg(repo, path, gh_root = "")
}
\arguments{
\item{repo}{Is the name of the repository.}

\item{path}{Is the path in the repo to the file.}

\item{gh_root}{If set to "" will try to access the file from your saved}
}
\description{
Function to get and set an rclone config file.
}
