% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_path}
\alias{get_data_path}
\title{Function to get the path to data.}
\usage{
get_data_path(
  cfg,
  protocol,
  task,
  kword,
  local_root = "",
  type = NA,
  data_path = NA
)
}
\description{
Function to get the path to data.
}
