% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_path_cfg}
\alias{get_data_path_cfg}
\title{Function to get the path to data given by cfg}
\usage{
get_data_path_cfg(cfg, protocol, kword, type = NA)
}
\description{
Function to get the path to data given by cfg
}
