% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_demo}
\alias{get_demo}
\title{Function to run the demographic stats.}
\usage{
get_demo(merged_df, defined_groups = NA)
}
\arguments{
\item{merged_df}{is a merged dataframe of Master Demo and a Project.}

\item{defined_groups}{a vector of strings identifying our groups.}
}
\value{
Returns a dataframe of the demographic data summary stats.
}
\description{
Function to run the demographic stats.
}
\examples{
get_demo(merged_df=my_df, defined_groups=c('HL', 'LL', 'HC', 'NON'))
}
