% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_explore2_clock_redcap}
\alias{get_explore2_clock_redcap}
\title{Function to filter explore2 scan data for clock, requires merge with explore.}
\usage{
get_explore2_clock_redcap(data, cfg, ...)
}
\description{
Function to filter explore2 scan data for clock, requires merge with explore.
}
