% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_explore_task_redcap}
\alias{get_explore_task_redcap}
\title{Function to filter the explore scan data by task.}
\usage{
get_explore_task_redcap(data, task, ...)
}
\description{
Function to filter the explore scan data by task.
}
