% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_github_creds}
\alias{get_github_creds}
\title{Function to grab the user's GitHub credentials}
\usage{
get_github_creds()
}
\value{
the credentials or NA if not found.
}
\description{
Tries to get the GitHub credentials from a small file under the path
'~/.github_creds.json'.
}
\examples{
set_github_creds()
}
