% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_id_path_grepl_count}
\alias{get_id_path_grepl_count}
\title{Function to get the number of matches if id paths to a grepl}
\usage{
get_id_path_grepl_count(full_path, requirement)
}
\description{
Function to get the number of matches if id paths to a grepl
}
