% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_masterdemo}
\alias{get_masterdemo}
\title{#' Convenience function to install Rclone
install_rclone <- function(module=FALSE) {
  # check that rclone is installed
  is_installed <- rclone.installed()
  # if rclone is not installed
  if(is_installed == FALSE) {
    # if doing a root user install
    if(module == FALSE) {
      install_str <- "curl https://rclone.org/install.sh | sudo bash"
        # otherwise, assume loading as a module
    } else {
      install_str <- "module load rclone"
    }
    # command to run on the system
    system(install_str, intern=TRUE)
  }
}
Function to create a REDCapR project from json input for master demographic.}
\usage{
get_masterdemo(cfg_path)
}
\arguments{
\item{cfg_path}{is the path to the lab's json configuration file.}
}
\value{
Returns an object to pull Master Demographic data with.
}
\description{
#' Convenience function to install Rclone
install_rclone <- function(module=FALSE) {
  # check that rclone is installed
  is_installed <- rclone.installed()
  # if rclone is not installed
  if(is_installed == FALSE) {
    # if doing a root user install
    if(module == FALSE) {
      install_str <- "curl https://rclone.org/install.sh | sudo bash"
        # otherwise, assume loading as a module
    } else {
      install_str <- "module load rclone"
    }
    # command to run on the system
    system(install_str, intern=TRUE)
  }
}
Function to create a REDCapR project from json input for master demographic.
}
\examples{
get_masterdemo(cfg_path='/bgfs/adombrovski/lab_resources/dnpl.json')
}
