% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_md_data}
\alias{get_md_data}
\title{Function to grab data from master demo and merge.}
\usage{
get_md_data(md, task_data, protocol, other_fields = c())
}
\arguments{
\item{md}{is a Master Demographic REDCapR object.}

\item{task_data}{the dataframe returned from REDCapR data object request.}

\item{protocol}{is the project we want data from.}

\item{other_fields}{a vector of extra fields to get from Master Demo.}
}
\value{
The selected Master Demo data merged with task/protocol data.
}
\description{
Function to grab data from master demo and merge.
}
\examples{
get_md_data(md=master_demo, task_data=bsocial_trust, protocol='bsocial')
}
