% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_merged_data}
\alias{get_merged_data}
\title{Function to grab data and merge it with master demo.}
\usage{
get_merged_data(cfg, protocol, task, other_fields = c())
}
\arguments{
\item{cfg}{is the path to the lab's json configuration file. (Required)}

\item{protocol}{is the project we want data from.}

\item{task}{is the specific task we want data for.}

\item{other_fields}{a vector of extra fields to get from Master Demo.}
}
\value{
The selected Master Demo data merged with task/protocol data.
}
\description{
Function to grab data and merge it with master demo.
}
\examples{
get_merged_data(md=master_demo, task_data=bsocial_trust, protocol='bsocial')
}
