% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_meson_data}
\alias{get_meson_data}
\title{Function to check for existence of legacy scanning data.}
\usage{
get_meson_data(cfg, protocol, task, local_root = "", data_path = NA)
}
\description{
Function to check for existence of legacy scanning data.
}
