% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_project}
\alias{get_project}
\title{Function to create a REDCapR project from json input.}
\usage{
get_project(cfg_path, protocol)
}
\arguments{
\item{cfg_path}{is the path to the lab's json configuration file.}

\item{protocol}{is the project we want data from.}
}
\value{
Returns an object to pull REDCap project data with.
}
\description{
Function to create a REDCapR project from json input.
}
\examples{
get_project(cfg_path='/bgfs/adombrovski/lab_resources/dnpl.json',
    protocol='bsocial')
}
