% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_real_id_regex}
\alias{get_real_id_regex}
\title{Function to get the regex to identify real participant ids of a protocol.}
\usage{
get_real_id_regex(cfg, protocol)
}
\description{
Function to get the regex to identify real participant ids of a protocol.
}
