% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_redcap_data}
\alias{get_redcap_data}
\title{Wrapper function to source and execute a function for grabbing task data
from REDCap.}
\usage{
get_redcap_data(data, protocol, task, fields = NA, events = NA, cfg = NA)
}
\arguments{
\item{data}{a REDCapR data object}

\item{protocol}{is the project we want data from.}

\item{task}{is the specific task we want data for.}

\item{fields}{are REDCap data fields we want to pull. (Default is all fields)}

\item{events}{are REDCap events we want to pull. (Default is all events)}

\item{cfg}{is the path to the lab's json configuration file. (Required)}
}
\value{
The selected REDCap data.
}
\description{
This funtion will first attempt to execute a default data fetching for the
protocol given that should be named:
  get_<protocol>_task_data()

  Example:
    get_bsocial_task_data()

If this general function fails, a specific one for that task will be used,
which should be name:
  get_<protocol>_<task>_data()

  Example:
    get_explore2_clock_data()

NOTE: This needs updated to allow selection from user input fields and
events instead of only those specified in the lab config file.
}
\examples{
get_redcap_data(data=bs, protocol='bsocial', task='trust', fields=cfg,
events=cfg, cfg='/bgfs/adombrovski/lab_resources/dnpl.json')
}
