% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_subj_ids_from_dir}
\alias{get_subj_ids_from_dir}
\title{Function to get the ids from a directory of subject data.}
\usage{
get_subj_ids_from_dir(cfg, protocol, data_path)
}
\description{
Function to get the ids from a directory of subject data.
}
