% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_task_completion_requirements}
\alias{get_task_completion_requirements}
\title{Function to get the required data for a task.}
\usage{
get_task_completion_requirements(cfg, task, modality = NA)
}
\description{
Function to get the required data for a task.
}
