% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_time_stamp}
\alias{get_time_stamp}
\title{Function to make a simple timestamp or return an empty string.}
\usage{
get_time_stamp(bool_in = TRUE)
}
\arguments{
\item{bool_in}{if set to FALSE, returns an empty string.}
}
\value{
a timestamp for the curret moment as a string.
}
\description{
Designed to return a timestamp as a string. Allows the input variable to
be set to an empty string so that you can always use this as part of a paste
command (in which an empty string will not modify the output string).
}
\examples{
get_time_stamp()
}
