% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{have_behavior_data}
\alias{have_behavior_data}
\title{Function to check for existence of behavioral data.
Assumes that data is stored in a directory identified as
<task>_<protocol>. This checks for a simple existence of the
subject ID as a directory under <task>_<protocol> and then
uses grep to check for the required files. This required at least
one match to each file. This will return a dataframe of the
subjects and whether or not they have complete behavioral data.}
\usage{
have_behavior_data(
  cfg,
  protocol,
  task,
  local_root = "",
  data_path = NA,
  my_required = NA
)
}
\description{
Function to check for existence of behavioral data.
Assumes that data is stored in a directory identified as
<task>_<protocol>. This checks for a simple existence of the
subject ID as a directory under <task>_<protocol> and then
uses grep to check for the required files. This required at least
one match to each file. This will return a dataframe of the
subjects and whether or not they have complete behavioral data.
}
