% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{have_bids_data}
\alias{have_bids_data}
\title{Function to check for existence of bids data.}
\usage{
have_bids_data(cfg, protocol, task, local_root = "", data_path = NA)
}
\description{
Function to check for existence of bids data.
}
