% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{have_data}
\alias{have_data}
\title{Wrapper function to check}
\usage{
have_data(
  cfg,
  modality,
  local_root = "",
  data_path = NA,
  my_required = NA,
  drop_failed = FALSE,
  ...
)
}
\arguments{
\item{cfg}{is the path to the lab's json configuration file. (Required)}

\item{modality}{is the modality of data to check (example: behavior)}

\item{local_root}{is the root directory to start checking from.}

\item{data_path}{is the path from the local_root to the data.}

\item{my_required}{is a subset of the requirements from cfg.}

\item{drop_failed}{will drop subjects without data if set to TRUE.}
}
\value{
Returns a dataframe of the modality of data requested.
}
\description{
This funtion will attempt to execute a check for the data of the given
modality that should be named:
  have_<modality>_data()

  Example:
    have_behavior_data()
}
\examples{
have_data(cfg='/Volumes/bierka_root/datamesh/behav/redcap3.json',
modality='behavior', protocol='bsocial', task='trust', local_root='/Volumes',
my_required = c("edat_scan", "text_scan"), drop_failed=TRUE)
}
