% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{have_dax_data}
\alias{have_dax_data}
\title{Function to check for existence of xnat scanning data.
Specifically, this will be data downloaded with the DNPL's
fork of DAX.}
\usage{
have_dax_data(cfg, protocol, task, local_root = "", data_path = NA)
}
\description{
Function to check for existence of xnat scanning data.
Specifically, this will be data downloaded with the DNPL's
fork of DAX.
}
