% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{have_daxmeson_data}
\alias{have_daxmeson_data}
\title{Function that checks for existence of xnat scanning data and then
checks meson data for anything not found.}
\usage{
have_daxmeson_data(
  cfg,
  protocol,
  task,
  local_root = "",
  xnat_path = NA,
  meson_path = NA
)
}
\description{
Function that checks for existence of xnat scanning data and then
checks meson data for anything not found.
}
