% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{have_meson_data}
\alias{have_meson_data}
\title{Below imaging data functions will probably have to be unique per protocol
and/or task.
^ Actually, probably not...seems to be formatted the same because of
downloading methods...
Function to check for existence of legacy scanning data.}
\usage{
have_meson_data(cfg, protocol, task, local_root = "", data_path = NA)
}
\description{
Below imaging data functions will probably have to be unique per protocol
and/or task.
^ Actually, probably not...seems to be formatted the same because of
downloading methods...
Function to check for existence of legacy scanning data.
}
