% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.mounted}
\alias{is.mounted}
\title{Function to check if the remote data is mounted.}
\usage{
is.mounted(mnt_path)
}
\arguments{
\item{mnt_path}{is the directory to check as a mount point.}
}
\value{
TRUE if the \code{mnt_point} already has a remote resource mounted.
FALSE if the \code{mnt_point} does not have a remote resource mounted.
}
\description{
Function to check if the remote data is mounted.
}
\examples{
is.mounted(mnt_path="/Users/bob/mnt")
}
