% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mnt_attempt_check}
\alias{mnt_attempt_check}
\title{Function to check that the mount succeeded or
increment attempt and run again.}
\usage{
mnt_attempt_check(
  mnt_path,
  remote_name,
  remote_path,
  attempt,
  max_attempts,
  trap,
  sleep
)
}
\description{
Function to check that the mount succeeded or
increment attempt and run again.
}
