% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mnt_remote_data}
\alias{mnt_remote_data}
\title{Function to mount remote data using rclone.}
\usage{
mnt_remote_data(
  mnt_path,
  remote_name,
  remote_path,
  attempt = 1,
  max_attempts = 5,
  trap = FALSE,
  sleep = 5
)
}
\arguments{
\item{mnt_path}{is the directory to try and remount.}

\item{remote_name}{is the name of the configured rclone remote.}

\item{remote_path}{is the path to the remote directory to mount locally.}

\item{attempt}{sets the attempt number you are on.}

\item{max_attempts}{sets the max number of mount attempts.}

\item{trap}{if set to TRUE, the mount will not persist after the session.}

\item{sleep}{number of seconds to sleep in between mount attempts.}
}
\description{
This can be used to mount SharePoint or any other cloud resource as if
the data existed on your local machine.
This uses rclone, ensure it is available.
rclone: https://rclone.org/commands/rclone_mount/
Mac OS Warning: While the files will mount and be accessible,
it has been seen where the Finder does not appropriately update
to display your remote folders and file. They do, however,
become accessible via terminal for normal operations like
cat, ls, cd, etc. Macs also require the deletion of .DS_Store folder
in order to attempt remounts as mount paths need to be empty.
}
\examples{
mnt_remote_data(mnt_path="/Users/bob/mnt", "Bob_OneDrive", "Documents")
}
