% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mount_Bierka}
\alias{mount_Bierka}
\title{Function for mounting the NAS, Bierka.}
\usage{
mount_Bierka()
}
\description{
This function is designed to mount Bierka to the user's home directory and
should only be used with the lab's default configuration file.
}
