% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mount_Skinner}
\alias{mount_Skinner}
\title{Function for mounting DNPLskinner SharePoint.}
\usage{
mount_Skinner()
}
\description{
This function is designed to mount Skinner to the user's home directory and
should only be used with the lab's default configuration file.
}
