% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{only.ds_store}
\alias{only.ds_store}
\title{Small function to check that .DS_Store is the only item in a directory.}
\usage{
only.ds_store(mnt_path)
}
\description{
Small function to check that .DS_Store is the only item in a directory.
}
