% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{os.is}
\alias{os.is}
\title{Small function to get OS string identifier.}
\usage{
os.is()
}
\value{
The system type as a string.
}
\description{
Small function to get OS string identifier.
}
\examples{
os.is()
}
