% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rclone.installed}
\alias{rclone.installed}
\title{Function to check that rclone is installed.}
\usage{
rclone.installed()
}
\value{
Boolean representing if rclone is installed or not.
}
\description{
Function to check that rclone is installed.
}
\examples{
rclone.installed()
}
