% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remnt_remote_data}
\alias{remnt_remote_data}
\title{Function to remount data: runs an unmount following by a mount.}
\usage{
remnt_remote_data(
  mnt_path,
  remote_name,
  remote_path,
  attempt = 1,
  max_attempts = 5
)
}
\arguments{
\item{mnt_path}{is the directory to try and remount.}

\item{remote_name}{is the name of the configured rclone remote.}

\item{remote_path}{is the path to the remote directory to mount locally.}
}
\description{
Function to remount data: runs an unmount following by a mount.
}
\examples{
remnt_remote_data(mnt_path="/Users/bob/mnt", "Bob_OneDrive", "Documents")
}
