% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_demographic_report}
\alias{run_demographic_report}
\title{Function to run the master demo data pull for a task/protocol.}
\usage{
run_demographic_report(
  cfg,
  protocol,
  task,
  other_fields = c(),
  out_dir = getwd(),
  load_env = FALSE,
  mnt_path = NA,
  remote_name = NA,
  remote_path = NA,
  max_attempts = 5,
  trap = FALSE,
  save = FALSE,
  include_timestamp = FALSE,
  return_data = FALSE,
  return_demo = FALSE,
  ...
)
}
\arguments{
\item{cfg}{is the path to the lab's json configuration file. (Required)}

\item{protocol}{is the project we want data from.}

\item{task}{is the specific task we want data for.}

\item{other_fields}{a vector of extra fields to get from Master Demo.}

\item{load_env}{if set to TRUE, will load the data to the global env.}

\item{mnt_path}{is the directory to try and remount.}

\item{remote_name}{is the name of the configured rclone remote.}

\item{remote_path}{is the path to the remote directory to mount locally.}

\item{max_attempts}{sets the max number of mount attempts.}

\item{trap}{if set to TRUE, the mount will not persist after the session.}

\item{save}{if set to TRUE, will save the results as csvs.}

\item{include_timestamp}{will add a timestamp to the saved output.}

\item{return_data}{if set to TRUE will return the resultant dataframe.}

\item{return_demo}{if set to TRUE will return the resultant demographic.}

\item{attempt}{sets the attempt number you are on.}
}
\value{
Based on return_data and return_demo inputs.
}
\description{
Function that can be used to mount remote data if needed, run a
demographic report, and chose how the result is saved. Will return
NA if return_data and return_demo are not TRUE. Otherwise will return a list
with either or both the data and demographic results.
}
\examples{
run_demographic_report(cfg="/Volumes/bierka_root/datamesh/behav/redcap3.json",
protocol="bsocial", task="trust", load_env=TRUE)
}
