% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_github_creds}
\alias{set_github_creds}
\title{Simple function to create/overwrite GitHub credentials}
\usage{
set_github_creds(username = NA, token = NA)
}
\arguments{
\item{username}{Your GitHub username.}

\item{token}{Your GitHub access token.}
}
\value{
FALSE if not set, TRUE if set.
}
\description{
To use this function:
First, create a GitHub account.
https://github.com/
Next, create a GitHub access token.
https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token
Finally, request access to the DNPL GitHub.
}
\examples{
set_github_creds(username=<username>, token=<token>)
}
