% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_lab.cfg}
\alias{set_lab.cfg}
\title{Function to set the lab config file.}
\usage{
set_lab.cfg(my_lab = NA, lab_cfg_path = NA)
}
\arguments{
\item{my_lab}{Is the name associated with the lab/project.}

\item{lab_cfg_path}{Is where you would like this new config to be saved.}
}
\description{
Sets the path to the lab's configuration file. Assumes the lab
is set to the global R environment variable "lab_name" defined
in this package. If setting up a new lab, override with by setting
my_lab. If you are using the same config as the global lab, it is
fine to just set my_lab. If you are creating a new lab config file,
use the lab_cfg_path variable.
}
