% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unmnt_remote_data}
\alias{unmnt_remote_data}
\title{Function to unmount remote data that was mounted via rclone.}
\usage{
unmnt_remote_data(mnt_path)
}
\arguments{
\item{mnt_path}{is the directory to end a mount on.}
}
\description{
Function to unmount remote data that was mounted via rclone.
}
\examples{
unmnt_remote_data(mnt_path="/Users/bob/mnt")
}
