% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_masterdemo_premerge}
\alias{validate_masterdemo_premerge}
\title{Function to try binding the master demo to the protocol data}
\usage{
validate_masterdemo_premerge(master_demo, protocol, protocol_name)
}
\description{
Function to try binding the master demo to the protocol data
}
