% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_masterdemo_updated}
\alias{validate_masterdemo_updated}
\title{Function to validate that master demo is up to date with the protocol}
\usage{
validate_masterdemo_updated(master_demo, protocol, protocol_name)
}
\description{
Function to validate that master demo is up to date with the protocol
}
