% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_masterdemo}
\alias{get_masterdemo}
\title{#' Convenience function to install Rclone
#' @return Boolean representing if rclone was successfully installed.
#' @examples
#' rclone.installed()
#' @export
install_rclone <- function() {
  # check that rclone is installed
  is_installed <- rclone.installed()
  # if rclone is not installed
  if(is_installed == FALSE) {
    # install string from rclone docs
    install_str <- "curl https://rclone.org/install.sh | sudo bash"
    # run the command on the system
    system(install_str, intern=TRUE)
    # check that rclone was successfully installed
  }
}
Function to create a REDCapR project from json input for master demographic.}
\usage{
get_masterdemo(cfg_path)
}
\arguments{
\item{cfg_path}{is the path to the lab's json configuration file.}
}
\value{
Returns an object to pull Master Demographic data with.
}
\description{
#' Convenience function to install Rclone
#' @return Boolean representing if rclone was successfully installed.
#' @examples
#' rclone.installed()
#' @export
install_rclone <- function() {
  # check that rclone is installed
  is_installed <- rclone.installed()
  # if rclone is not installed
  if(is_installed == FALSE) {
    # install string from rclone docs
    install_str <- "curl https://rclone.org/install.sh | sudo bash"
    # run the command on the system
    system(install_str, intern=TRUE)
    # check that rclone was successfully installed
  }
}
Function to create a REDCapR project from json input for master demographic.
}
\examples{
get_masterdemo(cfg_path='/bgfs/adombrovski/lab_resources/dnpl.json')
}
