% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lab.cfg}
\alias{lab.cfg}
\title{Small function to get the default lab config path}
\usage{
lab.cfg(cfg_env = NA)
}
\arguments{
\item{cfg_env}{is the environment variable to the config file.}
}
\value{
The lab config file path as a string.
}
\description{
Small function to get the default lab config path
}
\examples{
lab.cfg()
}
